refactor(plugin): extract shared dotenv lib (single source of truth)
Three copies of parseDotenv + readRembricSlug + SLUG_RE existed in
the JS/TS plugin tree:
  - plugin/bin/rembric-bridge.mjs                       (JS)
  - plugin/.opencode-plugin/plugin.ts                   (TS, inline)
  - plugin/.opencode-plugin/helpers.ts                  (TS, test mirror)

Replaced by a single canonical module plugin/bin/rembric-dotenv.mjs.
The bridge imports it via the relative path ./rembric-dotenv.mjs
(sibling file at runtime). The opencode plugin imports it via
../bin/rembric-dotenv.mjs at source time (resolves for `tsc --noEmit`
and `pnpm vitest` against the monorepo layout); install.sh sed-
substitutes that path with the absolute installed path
($HOME/.config/rembric/bin/rembric-dotenv.mjs) before copying
plugin.ts to the user's machine. Bun's ESM resolver in opencode
1.15.5 accepts absolute paths — confirmed against the live install.

helpers.ts is deleted. plugin.test.ts imports the helpers directly
from the shared lib for unit testing.

The previous byte-by-byte parity invariant test
(`plugin/.opencode-plugin/ helpers parity`) is replaced by a
static-grep invariant
(`plugin/bin/rembric-dotenv.mjs is the single source of truth for
slug parsing`) that fails the build if either plugin.ts or
rembric-bridge.mjs ever declares a local `function parseDotenv` or
`SLUG_RE = /` literal.

Bash (plugin/scripts/_api.sh) and Python
(plugin/.hermes-plugin/__init__.py) keep their own slug-parsing
implementations. Cross-language wrapping a 20-line dotenv parser
pays a per-call subprocess spawn that the duplication cost doesn't
justify; the regex value SHALL agree across implementations.

OpenSpec change `add-opencode-plugin` un-archived, amended with a
new requirement `Shared dotenv lib SHALL be the single source of
truth for slug parsing`, Decision 3b added to design.md, and
re-archived. Live specs at openspec/specs/opencode-plugin/spec.md
and openspec/specs/plugin-session-protocol/spec.md re-applied.

End-to-end re-verified against the dev stack: opencode 1.15.5
spawns the bridge, the bridge imports parseDotenv + SLUG_RE from
the sibling dotenv lib, builds /mcp/demo URL from .rembric, MCP
initialize returns toolCount=19, plugin.ts's session.created
handler POSTs the session row to /api/demo/sessions, sub-agent
session correctly suppressed by the parentID heuristic.

Validation: 464/464 tests pass, typecheck clean, lint clean,
`openspec validate add-opencode-plugin --strict` valid.
