+++ plugin/.opencode-plugin/uninstall.sh
@@ -5,25 +5,21 @@ set -uo pipefail
 
 PLUGIN_DEST="${HOME}/.config/opencode/plugins/rembric.ts"
 BRIDGE_DEST="${HOME}/.config/rembric/bin/rembric-bridge.mjs"
+DOTENV_DEST="${HOME}/.config/rembric/bin/rembric-dotenv.mjs"
 REMBRIC_BIN_DIR="${HOME}/.config/rembric/bin"
 REMBRIC_DIR="${HOME}/.config/rembric"
 
 removed=()
 absent=()
 
-if [ -e "${PLUGIN_DEST}" ]; then
-  rm -f "${PLUGIN_DEST}"
-  removed+=("${PLUGIN_DEST}")
-else
-  absent+=("${PLUGIN_DEST}")
-fi
-
-if [ -e "${BRIDGE_DEST}" ]; then
-  rm -f "${BRIDGE_DEST}"
-  removed+=("${BRIDGE_DEST}")
-else
-  absent+=("${BRIDGE_DEST}")
-fi
+for target in "${PLUGIN_DEST}" "${BRIDGE_DEST}" "${DOTENV_DEST}"; do
+  if [ -e "${target}" ]; then
+    rm -f "${target}"
+    removed+=("${target}")
+  else
+    absent+=("${target}")
+  fi
+done
 
 rmdir "${REMBRIC_BIN_DIR}" 2>/dev/null || true
 rmdir "${REMBRIC_DIR}" 2>/dev/null || true
